support updating foreign keys in PreparedUpdateQuery
